requirements.txt
gitpython
Sphinx
sphinx-autobuild
sphinx-rtd-theme
pandoc
